Type recv_buffers remain as uint32_t to match header payload sizes.
Avoids a useless size_t-to-uint64_t cast in the PAYLOAD_BAD_SIZE log on LP64.

[TEST COMMIT]

+++ src/cc/lib/swcdb/core/comm/ConnHandler.cc
@@ -423,7 +423,7 @@ void ConnHandler::read() noexcept {
 
 void ConnHandler::recv_buffers(Event::Ptr&& ev) {
   StaticBuffer* buffer;
-  size_t remain;
+  uint32_t remain;
   if(!ev->data.size) {
     buffer = &ev->data;
     remain = ev->header.data.size;
